+++ wepsim_i18n/i18n.js
@@ -106,10 +106,14 @@
 
     function i18n_get ( component, lang, key )
     {
+	if (typeof i18n.eltos[component] === "undefined") {
+	    return key ;
+	}
+
 	var translation = i18n.eltos[component][lang][key] ;
 
 	if (typeof translation === "undefined") {
-	    translation = key ;
+	    return key ;
 	}
 
 	return translation ;