+++ ws_travis/jshint.sh
@@ -6,14 +6,13 @@ set -x
  jshint ./sim_core/sim_api_native.js
 #jshint ./sim_core/sim_api_core.js
  jshint ./sim_core/sim_cfg.js
- jshint ./sim_core/sim_core_ui_console.js
+ jshint ./sim_core/sim_core_ui_debug.js
  jshint ./sim_core/sim_core_ui_notify.js
  jshint ./sim_core/sim_core_ui_help.js
  jshint ./sim_core/sim_core_ui_voice.js
  jshint ./sim_core/sim_core_ui_hw.js
- jshint ./sim_core/sim_core_ui_registers.js
+ jshint ./sim_core/sim_core_ui_details.js
  jshint ./sim_core/sim_core_breakpointicons.js
-#jshint ./sim_core/sim_core_ui_memory.js
 #jshint ./sim_core/sim_core_ctrl.js
 #jshint ./sim_core/sim_core_record.js